refactor: split synthetic client
